requirements.txt
python-instagram==0.8.0
flickrapi==1.4.2
